+++ utils/validation/contactValidationSchemas.js
@@ -1,0 +1,10 @@
+const Joi = require("joi");
+
+const contactValidationSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+  favorite: Joi.boolean(),
+});
+
+module.exports = contactValidationSchema;
