+++ utils/validation/index.js
@@ -1,0 +1,8 @@
+const contactValidationSchema = require("./contactValidationSchemas");
+
+const updateFavoriteSchema = require("./updateFavoriteSchema");
+
+module.exports = {
+  contactValidationSchema,
+  updateFavoriteSchema,
+};
